Replace custom language selector with default Google Translate widget
