+++ src/components/GoogleTranslateWidget.tsx
@@ -1,0 +1,30 @@
+import { useEffect } from 'react';
+
+const GoogleTranslateWidget = () => {
+  useEffect(() => {
+    if (document.getElementById('google-translate-script')) return;
+
+    window.googleTranslateElementInit = () => {
+      new window.google.translate.TranslateElement(
+        {
+          pageLanguage: 'en',
+          includedLanguages: 'en,hr,de',
+          autoDisplay: false,
+          layout: window.google.translate.TranslateElement.InlineLayout.SIMPLE,
+        },
+        'google_translate_element'
+      );
+    };
+
+    const script = document.createElement('script');
+    script.id = 'google-translate-script';
+    script.src =
+      '//translate.google.com/translate_a/element.js?cb=googleTranslateElementInit';
+    script.async = true;
+    document.body.appendChild(script);
+  }, []);
+
+  return <div id="google_translate_element" className="mt-2" />;
+};
+
+export default GoogleTranslateWidget;
